+++ spec/readymade/form/form_options_spec.rb
@@ -1,19 +1,20 @@
 # frozen_string_literal: true
+
 require 'byebug'
 
 RSpec.describe Readymade::Form::FormOptions do
   class TestForm < Readymade::Form
-    PERMITTED_ATTRIBUTES = %i[attr1 attr2 attr3]
-    REQUIRED_ATTRIBUTES = %i[attr1 attr3]
+    PERMITTED_ATTRIBUTES = %i[attr1 attr2 attr3].freeze
+    REQUIRED_ATTRIBUTES = %i[attr1 attr3].freeze
   end
 
   class TestFormWithFormOptions < Readymade::Form
-    PERMITTED_ATTRIBUTES = %i[attr1 attr2 attr3]
-    REQUIRED_ATTRIBUTES = %i[attr1 attr3]
+    PERMITTED_ATTRIBUTES = %i[attr1 attr2 attr3].freeze
+    REQUIRED_ATTRIBUTES = %i[attr1 attr3].freeze
 
     def form_options
       {
-        attr1: [1,2,3],
+        attr1: [1, 2, 3],
         passed_argument: args[:passed_argument]
       }
     end
@@ -21,7 +22,7 @@ RSpec.describe Readymade::Form::FormOptions do
 
   let(:args) do
     {
-      attr1: [1,2,3],
+      attr1: [1, 2, 3],
       attr2: 'attr two',
       attr3: 'attr three',
       passed_argument: 'passed_argument'
@@ -42,9 +43,9 @@ RSpec.describe Readymade::Form::FormOptions do
         instance = TestFormWithFormOptions.form_options(**args)
 
         expect(instance[:attr1]).to eq(args[:attr1])
-        expect(instance[:passed_argument]).to eq('passed_argument')
-        expect(instance[:attr2]).to eq(nil)
-        expect(instance[:attr3]).to eq(nil)
+        expect(instance[:passed_argument]).to be('passed_argument')
+        expect(instance[:attr2]).to be_nil
+        expect(instance[:attr3]).to be_nil
       end
     end
   end
@@ -53,9 +54,9 @@ RSpec.describe Readymade::Form::FormOptions do
     it 'returns boolean' do
       instance = TestForm.form_options(**args)
 
-      expect(instance.required?(:attr1)).to eq(true)
-      expect(instance.required?(:attr2)).to eq(false)
-      expect(instance.required?(:attr3)).to eq(true)
+      expect(instance.required?(:attr1)).to be(true)
+      expect(instance.required?(:attr2)).to be(false)
+      expect(instance.required?(:attr3)).to be(true)
     end
   end
 end