+++ spec/readymade_lead_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe Readymade do
   it 'has a version number' do
-    expect(Readymade::VERSION).not_to be nil
+    expect(Readymade::VERSION).not_to be_nil
   end
 end